requirements.txt
pandas>=1.3.0
numpy>=1.21.0
sqlite3>=3.35.0
psycopg2-binary>=2.9.0 